deleted the type adpater generator test, moved to TypeUtils test, since all the utility methods were moved
